+++ client/target.go
@@ -37,6 +37,13 @@ func (m *Target) APIServer() string {
 	return m.targetEntry.APIServer
 }
 
+// ShouldConfigureForGKE returns true iff the API server URL and CA
+// should be fetched from the kube-public ClientConfig provided by GKE clusters
+// instead of the other methods (e.g. inline in Osprey config file or from Osprey server)
+func (m *Target) ShouldConfigureForGKE() bool {
+	return m.targetEntry.UseGKEClientConfig
+}
+
 // ShouldFetchCAFromAPIServer returns true iff the CA should be fetched from the kube-public ConfigMap
 // instead of the other methods (e.g. inline in Osprey config file or from Osprey server)
 func (m *Target) ShouldFetchCAFromAPIServer() bool {